@@ -1,6 +1,9 @@
 package com.tanitim.universitetanitim;
 
-import android.support.v7.app.AppCompatActivity;
+
+
+import androidx.appcompat.app.AppCompatActivity;
+
 import com.jaeger.library.StatusBarUtil;
 
 
