@@ -7,25 +7,23 @@ import android.view.View;
 import android.view.ViewGroup;
 import android.widget.ImageView;
 import android.widget.TextView;
-
 import androidx.cardview.widget.CardView;
 import androidx.recyclerview.widget.RecyclerView;
-
 import com.bumptech.glide.Glide;
-import com.tanitim.universitetanitim.DetayActivity;
 import com.tanitim.universitetanitim.R;
 import com.tanitim.universitetanitim.Models.Universiteler;
 import com.tanitim.universitetanitim.retrofit.UniversitelerDAOinterface;
+import com.tanitim.universitetanitim.sehireGoreUniversite;
 
 import java.util.List;
 
 
-public class UniversitelerAdapter extends RecyclerView.Adapter<UniversitelerAdapter.CardTasarimTutucu> {
+public class SehirlerAdapter extends RecyclerView.Adapter<SehirlerAdapter.CardTasarimTutucu> {
     private Context mContext;
     private List<Universiteler> universitelerListe;
     private UniversitelerDAOinterface universitelerDAOinterface;
 
-    public UniversitelerAdapter(Context mContext, List<Universiteler> universitelerListe
+    public SehirlerAdapter(Context mContext, List<Universiteler> universitelerListe
             ,UniversitelerDAOinterface universitelerDAOinterface) {
         this.mContext = mContext;
         this.universitelerListe = universitelerListe;
@@ -34,7 +32,7 @@ public class UniversitelerAdapter extends RecyclerView.Adapter<UniversitelerAdap
 
     @Override
     public CardTasarimTutucu onCreateViewHolder(ViewGroup parent, int viewType) {
-        View view = LayoutInflater.from(parent.getContext()).inflate(R.layout.yenicard,parent,false);
+        View view = LayoutInflater.from(parent.getContext()).inflate(R.layout.card_tasarim,parent,false);
 
         return new CardTasarimTutucu(view);
     }
@@ -48,29 +46,36 @@ public class UniversitelerAdapter extends RecyclerView.Adapter<UniversitelerAdap
                 .getColor(R.color.purple_inactive), PorterDuff.Mode.SRC_IN); //ProgressBar Rengi */
 
 
-
-        Glide.with(mContext).load("https://tohere.net/yedek/logo/"+
-                universitelerListe.get(position).getSlug()+".png").into(holder.logo);   //Universite Logo
-        Glide.with(mContext).load("https://tohere.net/yedek/resim/"+
-                universitelerListe.get(position).getSlug()+".jpg").into(holder.arkaplan);//Universite resim
+        String il=karakterCevir(universitelerListe.get(position).getIl()); //türkçe karakter silmek için
+        Glide.with(mContext).load("https://tohere.net/yedek/sehir/"+
+               il+"_optimized"+".jpg").into(holder.arkaplan);//Universite resim
 
 
 
-        holder.textViewIsim.setText(universite.getIsim()); //universite adı textview
-        holder.textViewDetay.setText(universite.getIl()); // il textView
+        holder.textViewSehir.setText(universite.getIl()); //universite adı textview
+        // il textView
 
         //carda tıklandığında olucaklar
-        holder.universiteler_card.setOnClickListener(new View.OnClickListener() {
+        holder.sehirler_card.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View view) {
-                Intent intent = new Intent(mContext, DetayActivity.class);
-
+                Intent intent = new Intent(mContext, sehireGoreUniversite.class);
                 intent.putExtra("nesne",  universite);
-
                 mContext.startActivity(intent);
             }
         });
 
+    }
+    public static String karakterCevir(String kelime)
+    {
+        String mesaj = kelime;
+        char[] oldValue = new char[] { 'ö', 'Ö', 'ü', 'Ü', 'ç', 'Ç', 'İ', 'ı', 'Ğ', 'ğ', 'Ş', 'ş' };
+        char[] newValue = new char[] { 'o', 'O', 'u', 'U', 'c', 'C', 'I', 'i', 'G', 'g', 'S', 's' };
+        for (int sayac = 0; sayac < oldValue.length; sayac++)
+        {
+            mesaj = mesaj.replace(oldValue[sayac], newValue[sayac]);
+        }
+        return mesaj;
     }
 
     @Override
@@ -81,18 +86,16 @@ public class UniversitelerAdapter extends RecyclerView.Adapter<UniversitelerAdap
     }
 
     public class CardTasarimTutucu extends RecyclerView.ViewHolder{
-        private TextView textViewIsim;
-        private TextView textViewDetay;
-        private CardView universiteler_card;
-        private ImageView logo,arkaplan;
+        private TextView textViewSehir;
+
+        private CardView sehirler_card;
+        private ImageView arkaplan;
 
 
         public CardTasarimTutucu(View itemView) {
             super(itemView);
-            textViewIsim = itemView.findViewById(R.id.textViewIsim);
-            textViewDetay = itemView.findViewById(R.id.textViewDetay);
-            universiteler_card = itemView.findViewById(R.id.universiteler_card);
-            logo = itemView.findViewById(R.id.logo);
+            textViewSehir = itemView.findViewById(R.id.textViewSehir);
+            sehirler_card = itemView.findViewById(R.id.sehirler_card);
             arkaplan = itemView.findViewById(R.id.arkaplan);
 
         }
