@@ -2,10 +2,12 @@ package com.tanitim.universitetanitim.Fragments;
 
 
 import android.os.Bundle;
-import android.support.v4.app.Fragment;
+
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
+
+import androidx.fragment.app.Fragment;
 
 import com.tanitim.universitetanitim.R;
 
