@@ -1,182 +1,109 @@
+
 package com.tanitim.universitetanitim.Models;
+
+import com.google.gson.annotations.Expose;
+import com.google.gson.annotations.SerializedName;
 
 import java.io.Serializable;
 
-
-
 public class Universiteler implements Serializable {
-    private String adres;
-    private int arastirma;
-    private int docent;
-    private int doktor;
-    private int doktoraerkek;
-    private int doktorakadin;
-    private int doktoratoplam;
-    private String eposta;
-    private String il;
+
+    @SerializedName("isim")
+    @Expose
     private String isim;
-    private String kurulus;
-    private int lisanserkek;
-    private int lisanskadin;
-    private int lisanstoplam;
-    private int ogretim;
-    private int onlisanserkek;
-    private int onlisanskadin;
-    private int onlisanstoplam;
-    private int profesor;
-    private String rektor;
+    @SerializedName("slug")
+    @Expose
     private String slug;
-    private int toplam;
-    private int toplamerkek;
-    private int toplamkadin;
+    @SerializedName("kurulus")
+    @Expose
+    private String kurulus;
+    @SerializedName("tur")
+    @Expose
     private String tur;
+    @SerializedName("il")
+    @Expose
+    private String il;
+    @SerializedName("website")
+    @Expose
     private String website;
-    private int yukseklisanserkek;
-    private int yukseklisanskadin;
-    private int yukseklisanstoplam;
-
+    @SerializedName("eposta")
+    @Expose
+    private String eposta;
+    @SerializedName("fax")
+    @Expose
     private String fax;
-    private String bolge;
+    @SerializedName("adres")
+    @Expose
+    private String adres;
+    @SerializedName("rektor")
+    @Expose
+    private String rektor;
+    @SerializedName("profesor")
+    @Expose
+    private String profesor;
+    @SerializedName("docent")
+    @Expose
+    private String docent;
+    @SerializedName("doktor")
+    @Expose
+    private String doktor;
+    @SerializedName("ogretim")
+    @Expose
+    private String ogretim;
+    @SerializedName("arastirma")
+    @Expose
+    private String arastirma;
+    @SerializedName("onlisanserkek")
+    @Expose
+    private String onlisanserkek;
+    @SerializedName("onlisanskadin")
+    @Expose
+    private String onlisanskadin;
+    @SerializedName("onlisanstoplam")
+    @Expose
+    private String onlisanstoplam;
+    @SerializedName("telefon")
+    @Expose
     private String telefon;
-
-    public Universiteler(String adres, int arastirma, int docent, int doktor, int doktoraerkek, int doktorakadin, int doktoratoplam, String eposta, String il, String isim, String kurulus, int lisanserkek, int lisanskadin, int lisanstoplam, int ogretim, int onlisanserkek, int onlisanskadin, int onlisanstoplam, int profesor, String rektor, String slug, int toplam, int toplamerkek, int toplamkadin, String tur, String website, int yukseklisanserkek, int yukseklisanskadin, int yukseklisanstoplam, String fax, String bolge, String telefon) {
-        this.adres = adres;
-        this.arastirma = arastirma;
-        this.docent = docent;
-        this.doktor = doktor;
-        this.doktoraerkek = doktoraerkek;
-        this.doktorakadin = doktorakadin;
-        this.doktoratoplam = doktoratoplam;
-        this.eposta = eposta;
-        this.il = il;
-        this.isim = isim;
-        this.kurulus = kurulus;
-        this.lisanserkek = lisanserkek;
-        this.lisanskadin = lisanskadin;
-        this.lisanstoplam = lisanstoplam;
-        this.ogretim = ogretim;
-        this.onlisanserkek = onlisanserkek;
-        this.onlisanskadin = onlisanskadin;
-        this.onlisanstoplam = onlisanstoplam;
-        this.profesor = profesor;
-        this.rektor = rektor;
-        this.slug = slug;
-        this.toplam = toplam;
-        this.toplamerkek = toplamerkek;
-        this.toplamkadin = toplamkadin;
-        this.tur = tur;
-        this.website = website;
-        this.yukseklisanserkek = yukseklisanserkek;
-        this.yukseklisanskadin = yukseklisanskadin;
-        this.yukseklisanstoplam = yukseklisanstoplam;
-
-        this.fax = fax;
-        this.bolge = bolge;
-        this.telefon = telefon;
-    }
-    public Universiteler(String il){
-        this.il = il;
-
-    }
-
-    public Universiteler() {
-    }
-
-    public String getFax() {
-        return fax;
-    }
-
-    public void setFax(String fax) {
-        this.fax = fax;
-    }
-
-    public String getBolge() {
-        return bolge;
-    }
-
-    public void setBolge(String bolge) {
-        this.bolge = bolge;
-    }
-
-    public String getTelefon() {
-        return telefon;
-    }
-
-    public void setTelefon(String telefon) {
-        this.telefon = telefon;
-    }
-
-    public String getAdres() {
-        return adres;
-    }
-
-    public void setAdres(String adres) {
-        this.adres = adres;
-    }
-
-    public int getArastirma() {
-        return arastirma;
-    }
-
-    public void setArastirma(int arastirma) {
-        this.arastirma = arastirma;
-    }
-
-    public int getDocent() {
-        return docent;
-    }
-
-    public void setDocent(int docent) {
-        this.docent = docent;
-    }
-
-    public int getDoktor() {
-        return doktor;
-    }
-
-    public void setDoktor(int doktor) {
-        this.doktor = doktor;
-    }
-
-    public int getDoktoraerkek() {
-        return doktoraerkek;
-    }
-
-    public void setDoktoraerkek(int doktoraerkek) {
-        this.doktoraerkek = doktoraerkek;
-    }
-
-    public int getDoktorakadin() {
-        return doktorakadin;
-    }
-
-    public void setDoktorakadin(int doktorakadin) {
-        this.doktorakadin = doktorakadin;
-    }
-
-    public int getDoktoratoplam() {
-        return doktoratoplam;
-    }
-
-    public void setDoktoratoplam(int doktoratoplam) {
-        this.doktoratoplam = doktoratoplam;
-    }
-
-    public String getEposta() {
-        return eposta;
-    }
-
-    public void setEposta(String eposta) {
-        this.eposta = eposta;
-    }
-
-    public String getIl() {
-        return il;
-    }
-
-    public void setIl(String il) {
-        this.il = il;
-    }
+    @SerializedName("bolge")
+    @Expose
+    private String bolge;
+    @SerializedName("lisanserkek")
+    @Expose
+    private String lisanserkek;
+    @SerializedName("lisanskadin")
+    @Expose
+    private String lisanskadin;
+    @SerializedName("lisanstoplam")
+    @Expose
+    private String lisanstoplam;
+    @SerializedName("yukseklisanserkek")
+    @Expose
+    private String yukseklisanserkek;
+    @SerializedName("yukseklisanskadin")
+    @Expose
+    private String yukseklisanskadin;
+    @SerializedName("yukseklisanstoplam")
+    @Expose
+    private String yukseklisanstoplam;
+    @SerializedName("doktoraerkek")
+    @Expose
+    private String doktoraerkek;
+    @SerializedName("doktorakadin")
+    @Expose
+    private String doktorakadin;
+    @SerializedName("doktoratoplam")
+    @Expose
+    private String doktoratoplam;
+    @SerializedName("toplamerkek")
+    @Expose
+    private String toplamerkek;
+    @SerializedName("toplamkadin")
+    @Expose
+    private String toplamkadin;
+    @SerializedName("toplam")
+    @Expose
+    private String toplam;
 
     public String getIsim() {
         return isim;
@@ -184,86 +111,6 @@ public class Universiteler implements Serializable {
 
     public void setIsim(String isim) {
         this.isim = isim;
-    }
-
-    public String getKurulus() {
-        return kurulus;
-    }
-
-    public void setKurulus(String kurulus) {
-        this.kurulus = kurulus;
-    }
-
-    public int getLisanserkek() {
-        return lisanserkek;
-    }
-
-    public void setLisanserkek(int lisanserkek) {
-        this.lisanserkek = lisanserkek;
-    }
-
-    public int getLisanskadin() {
-        return lisanskadin;
-    }
-
-    public void setLisanskadin(int lisanskadin) {
-        this.lisanskadin = lisanskadin;
-    }
-
-    public int getLisanstoplam() {
-        return lisanstoplam;
-    }
-
-    public void setLisanstoplam(int lisanstoplam) {
-        this.lisanstoplam = lisanstoplam;
-    }
-
-    public int getOgretim() {
-        return ogretim;
-    }
-
-    public void setOgretim(int ogretim) {
-        this.ogretim = ogretim;
-    }
-
-    public int getOnlisanserkek() {
-        return onlisanserkek;
-    }
-
-    public void setOnlisanserkek(int onlisanserkek) {
-        this.onlisanserkek = onlisanserkek;
-    }
-
-    public int getOnlisanskadin() {
-        return onlisanskadin;
-    }
-
-    public void setOnlisanskadin(int onlisanskadin) {
-        this.onlisanskadin = onlisanskadin;
-    }
-
-    public int getOnlisanstoplam() {
-        return onlisanstoplam;
-    }
-
-    public void setOnlisanstoplam(int onlisanstoplam) {
-        this.onlisanstoplam = onlisanstoplam;
-    }
-
-    public int getProfesor() {
-        return profesor;
-    }
-
-    public void setProfesor(int profesor) {
-        this.profesor = profesor;
-    }
-
-    public String getRektor() {
-        return rektor;
-    }
-
-    public void setRektor(String rektor) {
-        this.rektor = rektor;
     }
 
     public String getSlug() {
@@ -274,28 +121,12 @@ public class Universiteler implements Serializable {
         this.slug = slug;
     }
 
-    public int getToplam() {
-        return toplam;
+    public String getKurulus() {
+        return kurulus;
     }
 
-    public void setToplam(int toplam) {
-        this.toplam = toplam;
-    }
-
-    public int getToplamerkek() {
-        return toplamerkek;
-    }
-
-    public void setToplamerkek(int toplamerkek) {
-        this.toplamerkek = toplamerkek;
-    }
-
-    public int getToplamkadin() {
-        return toplamkadin;
-    }
-
-    public void setToplamkadin(int toplamkadin) {
-        this.toplamkadin = toplamkadin;
+    public void setKurulus(String kurulus) {
+        this.kurulus = kurulus;
     }
 
     public String getTur() {
@@ -306,6 +137,14 @@ public class Universiteler implements Serializable {
         this.tur = tur;
     }
 
+    public String getIl() {
+        return il;
+    }
+
+    public void setIl(String il) {
+        this.il = il;
+    }
+
     public String getWebsite() {
         return website;
     }
@@ -314,29 +153,212 @@ public class Universiteler implements Serializable {
         this.website = website;
     }
 
-    public int getYukseklisanserkek() {
+    public String getEposta() {
+        return eposta;
+    }
+
+    public void setEposta(String eposta) {
+        this.eposta = eposta;
+    }
+
+    public String getFax() {
+        return fax;
+    }
+
+    public void setFax(String fax) {
+        this.fax = fax;
+    }
+
+    public String getAdres() {
+        return adres;
+    }
+
+    public void setAdres(String adres) {
+        this.adres = adres;
+    }
+
+    public String getRektor() {
+        return rektor;
+    }
+
+    public void setRektor(String rektor) {
+        this.rektor = rektor;
+    }
+
+    public String getProfesor() {
+        return profesor;
+    }
+
+    public void setProfesor(String profesor) {
+        this.profesor = profesor;
+    }
+
+    public String getDocent() {
+        return docent;
+    }
+
+    public void setDocent(String docent) {
+        this.docent = docent;
+    }
+
+    public String getDoktor() {
+        return doktor;
+    }
+
+    public void setDoktor(String doktor) {
+        this.doktor = doktor;
+    }
+
+    public String getOgretim() {
+        return ogretim;
+    }
+
+    public void setOgretim(String ogretim) {
+        this.ogretim = ogretim;
+    }
+
+    public String getArastirma() {
+        return arastirma;
+    }
+
+    public void setArastirma(String arastirma) {
+        this.arastirma = arastirma;
+    }
+
+    public String getOnlisanserkek() {
+        return onlisanserkek;
+    }
+
+    public void setOnlisanserkek(String onlisanserkek) {
+        this.onlisanserkek = onlisanserkek;
+    }
+
+    public String getOnlisanskadin() {
+        return onlisanskadin;
+    }
+
+    public void setOnlisanskadin(String onlisanskadin) {
+        this.onlisanskadin = onlisanskadin;
+    }
+
+    public String getOnlisanstoplam() {
+        return onlisanstoplam;
+    }
+
+    public void setOnlisanstoplam(String onlisanstoplam) {
+        this.onlisanstoplam = onlisanstoplam;
+    }
+
+    public String getTelefon() {
+        return telefon;
+    }
+
+    public void setTelefon(String telefon) {
+        this.telefon = telefon;
+    }
+
+    public String getBolge() {
+        return bolge;
+    }
+
+    public void setBolge(String bolge) {
+        this.bolge = bolge;
+    }
+
+    public String getLisanserkek() {
+        return lisanserkek;
+    }
+
+    public void setLisanserkek(String lisanserkek) {
+        this.lisanserkek = lisanserkek;
+    }
+
+    public String getLisanskadin() {
+        return lisanskadin;
+    }
+
+    public void setLisanskadin(String lisanskadin) {
+        this.lisanskadin = lisanskadin;
+    }
+
+    public String getLisanstoplam() {
+        return lisanstoplam;
+    }
+
+    public void setLisanstoplam(String lisanstoplam) {
+        this.lisanstoplam = lisanstoplam;
+    }
+
+    public String getYukseklisanserkek() {
         return yukseklisanserkek;
     }
 
-    public void setYukseklisanserkek(int yukseklisanserkek) {
+    public void setYukseklisanserkek(String yukseklisanserkek) {
         this.yukseklisanserkek = yukseklisanserkek;
     }
 
-    public int getYukseklisanskadin() {
+    public String getYukseklisanskadin() {
         return yukseklisanskadin;
     }
 
-    public void setYukseklisanskadin(int yukseklisanskadin) {
+    public void setYukseklisanskadin(String yukseklisanskadin) {
         this.yukseklisanskadin = yukseklisanskadin;
     }
 
-    public int getYukseklisanstoplam() {
+    public String getYukseklisanstoplam() {
         return yukseklisanstoplam;
     }
 
-    public void setYukseklisanstoplam(int yukseklisanstoplam) {
+    public void setYukseklisanstoplam(String yukseklisanstoplam) {
         this.yukseklisanstoplam = yukseklisanstoplam;
     }
 
+    public String getDoktoraerkek() {
+        return doktoraerkek;
+    }
+
+    public void setDoktoraerkek(String doktoraerkek) {
+        this.doktoraerkek = doktoraerkek;
+    }
+
+    public String getDoktorakadin() {
+        return doktorakadin;
+    }
+
+    public void setDoktorakadin(String doktorakadin) {
+        this.doktorakadin = doktorakadin;
+    }
+
+    public String getDoktoratoplam() {
+        return doktoratoplam;
+    }
+
+    public void setDoktoratoplam(String doktoratoplam) {
+        this.doktoratoplam = doktoratoplam;
+    }
+
+    public String getToplamerkek() {
+        return toplamerkek;
+    }
+
+    public void setToplamerkek(String toplamerkek) {
+        this.toplamerkek = toplamerkek;
+    }
+
+    public String getToplamkadin() {
+        return toplamkadin;
+    }
+
+    public void setToplamkadin(String toplamkadin) {
+        this.toplamkadin = toplamkadin;
+    }
+
+    public String getToplam() {
+        return toplam;
+    }
+
+    public void setToplam(String toplam) {
+        this.toplam = toplam;
+    }
 
 }
